Move CommonFields abstract model to organisation/models.py.
